+++ test/components/HomeSpec.js
@@ -11,37 +11,22 @@ import setup from '../setup.js'
 // Data
 import data from '../data/projects.json'
 import { getInitialState } from '../../src/getInitialState'
+import mapStateToProps from '../../src/containers/HomePage/mapStateToProps'
 
 setup()
 
 // Components to check
 import Home from '../../src/components/home/Home'
-import ProjectList from '../../src/components/projects/ProjectTable'
+import ProjectList from '../../src/components/projects/ProjectList'
+import ProjectCard from '../../src/components/projects/ProjectCard'
 
 import populate from '../../src/helpers/populate'
 
-function getHotProjects(state) {
+function getHotProjects (state) {
   return state.githubProjects.daily
     .map(id => state.entities.projects[id])
     .slice(0, 20)
     .map(populate(state.entities.tags))
-}
-
-function mapStateToProps(state) {
-  const {
-    entities: { projects, tags },
-    githubProjects
-  } = state
-
-  const popularProjects = githubProjects.total
-    .map(id => projects[id])
-    .slice(0, 20)
-    .map(populate(tags))
-
-  return {
-    hotProjects: getHotProjects(state),
-    popularProjects
-  }
 }
 
 test('Check <ProjectList>', assert => {
@@ -50,13 +35,12 @@ test('Check <ProjectList>', assert => {
 
   const component = shallow(
     <ProjectList
-      projects = { getHotProjects(state) }
+      projects={getHotProjects(state)}
     />
   )
-  assert.ok(component, `The component should exist.`)
-  // console.log(component.debug())
-  const foundProjects = component.find(ProjectList.Row)
-  assert.equal(foundProjects.length, count, `There should be N projects.`)
+  assert.ok(component, 'The component should exist.')
+  const foundProjects = component.find(ProjectCard)
+  assert.equal(foundProjects.length, count, 'There should be N projects.')
   assert.end()
 })
 
@@ -67,18 +51,19 @@ test('Check <Home> component', (assert) => {
 
   const component = mount(
     <Home
-      hotProjects={props.hotProjects }
+      hotProjects={props.hotProjects}
       popularProjects={props.popularProjects}
       uiActions={{}}
       authActions={{}}
+      popularTags={props.popularTags}
     />
   )
 
-  assert.ok(component, `The component should exist.`)
+  assert.ok(component, 'The component should exist.')
 
   const lists = component.find(ProjectList)
 
-  assert.equal(lists.length, 2, `There should be 2 lists of projects.`)
+  assert.equal(lists.length, 1, 'There should be 1 lists of projects.')
 
   assert.end()
 })